Remove primitives::constants module
Inline the code out of `primitives/constants.rs` into
`primitives/mod.rs`.

Refactor only, no logic changes.
