+++ bitcoin/src/primitives/block.rs
@@ -22,7 +22,7 @@ use crate::util::uint::Uint256;
 use crate::consensus::{encode, Encodable, Decodable};
 use crate::network::constants::Network;
 use crate::primitives::transaction::Transaction;
-use crate::primitives::constants::{max_target, WITNESS_SCALE_FACTOR};
+use crate::primitives::{max_target, WITNESS_SCALE_FACTOR};
 use crate::script;
 use crate::VarInt;
 use crate::internal_macros::impl_consensus_encoding;
