+++ bitcoin/src/util/merkleblock.rs
@@ -50,7 +50,7 @@ use crate::hashes::Hash;
 use crate::hash_types::{Txid, TxMerkleNode};
 
 use crate::primitives::transaction::Transaction;
-use crate::primitives::constants::{MAX_BLOCK_WEIGHT, MIN_TRANSACTION_WEIGHT};
+use crate::primitives::{MAX_BLOCK_WEIGHT, MIN_TRANSACTION_WEIGHT};
 use crate::consensus::encode::{self, Decodable, Encodable};
 use crate::util::merkleblock::MerkleBlockError::*;
 use crate::{Block, BlockHeader};
